Solve the issue of spazzy teleports
QR Teleporters that are two-way were having the issue of constantly
teleporting the player between the two posts. This solves the issue with
a level-wide Teleport Lock as a timer. When the player enters a QR
Teleporter, the timer runs for half a second (as of writing this) and
then releases. Of course, this has to be checked for in code, so this is
only a soft-ish teleport lock. Hey, it works though!